Move methods from BaseDsiExporter to DFESignIn
This is because I expect to need these methods for UpdateDfeSignInUsers.
