+++ lib/base_dsi_exporter.rb
@@ -2,6 +2,8 @@ require 'dfe_sign_in_api'
 require 'google/cloud/bigquery'
 
 class BaseDsiExporter
+  include DFESignIn
+
   attr_reader :dataset
 
   def initialize(bigquery: Google::Cloud::Bigquery.new)
@@ -23,20 +25,5 @@ class BaseDsiExporter
 
       insert_table_data(response['users'])
     end
-  end
-
-  def number_of_pages
-    response = api_response
-    raise (response['message'] || 'failed request') if response['numberOfPages'].nil?
-
-    response['numberOfPages']
-  end
-
-  def users_nil_or_empty?(response)
-    response['users'].nil? || response['users'].first.empty?
-  end
-
-  def error_message_for(response)
-    response['message'] || 'failed request'
   end
 end
