+++ lib/update_dsi_users_in_db.rb
@@ -1,2 +1,8 @@
+require 'dfe_sign_in_api'
+
 class UpdateDfeSignInUsers
+
+    def run
+    end
+
 end
